# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Orvex Wiki Engine — Orvex Primitive Surface
  version: 0.1.0-draft
  summary: Additive /api/orvex/* primitives layered on the AGPL Docmost engine.
  description: >-
    The ADDITIVE orvex primitive surface of the AGPL wiki engine — the atomic
    write chokepoint (FR-W1), the F-QUOTA usage-vs-caps read (FR-W15), the
    identity exchange-token → engine-session mint (FR-W6), the AGPL §13 source
    offer (FR-W19), and the A-MOVE tenant-move step contract. It does NOT
    re-describe the upstream Docmost REST API; the public programmatic front is
    `wiki-api` (D-S16/D-S18).


    ## Audience / tier (INTERNAL)

    This API is INTERNAL. It is consumed ONLY by the first-party React client
    and by `orvex-wiki-api` (the per-cell satellite that fronts the engine for
    the Studio MCP gateway). It is NEVER publicly agent-facing — agents reach the
    engine through `wiki-api`, never here. See `servers[].x-tier`.


    ## Genuinely-real health surface (upstream, terminus-backed) — NOT re-authored here

    The engine already exposes a REAL health surface from upstream Docmost. It is
    documented here for provenance only (it is not an orvex primitive):

    - `GET /api/health` — `@nestjs/terminus` `HealthCheck`, pings Postgres +
      Redis. Live shape (SkipTransform, raw terminus):
      `{"status":"ok","info":{"database":{"status":"up"},"redis":{"status":"up"}},"error":{},"details":{...}}`.
...

...
    See `x-real-health-surface`.


...

...
  license:
...

...

...
servers:
...
security:
  - engineSession: []
...
paths:
...
  /api/orvex/tenant-move/quiesce:
    post:
      tags: [tenant-move]
      operationId: orvexTenantMoveQuiesce
      summary: Tenant-move step 1 — freeze writes for the tenant on the source cell
      x-status: draft
      x-classification: noop-501
      description: >-
        A-MOVE step 1/4 (quiesce). Freeze writes for the tenant on the SOURCE cell.
        The TYPED CONTRACT is the non-retrofittable, day-1 part; the body is
        deliberately a 501 stub — A-MOVE / cell-contract rule #10: "the 501 stub is
        5 lines any day, the typed contract is not". Request/response reference the
        TenantMoveManifest (field naming mirrors
        `../orvex-studio-contracts/schemas/org-move-manifest.schema.json`).
        `Idempotency-Key` is MANDATORY (cell-contract rule #11): 421-retries,
        Stripe retries and Temporal retries all replay these calls.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/TenantMoveManifestBody'
      responses:
        '200':
          $ref: '#/components/responses/TenantMoveStepAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          description: TODAY — typed day-1 contract, deliberately 501. Typed sentinel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedError'
              example:
                code: NOT_IMPLEMENTED
                operationId: orvexTenantMoveQuiesce
                marker: ORVEX_NOT_IMPLEMENTED
  /api/orvex/tenant-move/export:
    post:
      tags: [tenant-move]
      operationId: orvexTenantMoveExport
      summary: Tenant-move step 2 — export the tenant's stores / S3 prefixes / cursors
      x-status: draft
      x-classification: noop-501
      description: >-
        A-MOVE step 2/4 (export). Dump the tenant's stores / S3 prefixes / cursors
        per the TenantMoveManifest. Typed day-1 contract, deliberately 501 (rule
        #10). `Idempotency-Key` MANDATORY (rule #11). See the OPEN store-list
        decision on the TenantMoveManifest schema.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/TenantMoveManifestBody'
      responses:
        '200':
          $ref: '#/components/responses/TenantMoveStepAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          description: TODAY — typed day-1 contract, deliberately 501. Typed sentinel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedError'
              example:
                code: NOT_IMPLEMENTED
                operationId: orvexTenantMoveExport
                marker: ORVEX_NOT_IMPLEMENTED
  /api/orvex/tenant-move/import:
    post:
      tags: [tenant-move]
      operationId: orvexTenantMoveImport
      summary: Tenant-move step 3 — load the tenant's stores into the destination cell
      x-status: draft
      x-classification: noop-501
      description: >-
        A-MOVE step 3/4 (import). Load the exported stores / S3 prefixes / cursors
        into the DESTINATION cell per the TenantMoveManifest. Typed day-1 contract,
        deliberately 501 (rule #10). `Idempotency-Key` MANDATORY (rule #11).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/TenantMoveManifestBody'
      responses:
        '200':
          $ref: '#/components/responses/TenantMoveStepAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          description: TODAY — typed day-1 contract, deliberately 501. Typed sentinel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedError'
              example:
                code: NOT_IMPLEMENTED
                operationId: orvexTenantMoveImport
                marker: ORVEX_NOT_IMPLEMENTED
  /api/orvex/tenant-move/activate:
    post:
      tags: [tenant-move]
      operationId: orvexTenantMoveActivate
      summary: Tenant-move step 4 — flip the registry tenant→cell mapping and resume writes
      x-status: draft
      x-classification: noop-501
      description: >-
        A-MOVE step 4/4 (activate). Flip the registry `tenant→cell` mapping (with a
        fresh `cell_epoch`) and resume writes on the DESTINATION cell. Typed day-1
        contract, deliberately 501 (rule #10). `Idempotency-Key` MANDATORY (rule
        #11) — a retried activate with a seen key returns the original result,
        never a double-flip.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/TenantMoveManifestBody'
      responses:
        '200':
          $ref: '#/components/responses/TenantMoveStepAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          description: TODAY — typed day-1 contract, deliberately 501. Typed sentinel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedError'
              example:
                code: NOT_IMPLEMENTED
                operationId: orvexTenantMoveActivate
                marker: ORVEX_NOT_IMPLEMENTED
components:
  securitySchemes:
    engineSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Engine session — the identity-minted exchange token folded into the
        upstream Docmost session (FR-W6). Deny-by-default. Presented by the React
        client and by orvex-wiki-api on the internal seam. The session-establishing
        endpoint (POST /api/orvex/session/exchange) is itself public.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        MANDATORY on every tenant-move step (cell-contract rule #11). A retried
        call carrying a seen key returns the ORIGINAL result, never a double-apply.
      schema:
        type: string
        minLength: 1
  requestBodies:
    TenantMoveManifestBody:
      required: true
      description: >-
        The versioned tenant-move manifest. Field naming mirrors
        ../orvex-studio-contracts/schemas/org-move-manifest.schema.json.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TenantMoveManifest'
  responses:
    Unauthorized:
      description: Missing/invalid engine session (deny-by-default). Framework Nest shape (observed live).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NestError'
          example:
            statusCode: 401
            error: Unauthorized
            message: Unauthorized
...
    TenantMoveStepAccepted:
      description: >-
        INTENDED — the step was accepted / applied (delivery fills the receipt
        shape in). Coarse-but-real: echoes the tenant + a per-step receipt.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TenantMoveStepResult'
  schemas:
...
    NotImplementedError:
      type: object
      additionalProperties: false
      required: [code, operationId, marker]
      description: >-
        Typed NotImplemented body returned by every 501 orvex primitive. NEVER a
        plausible-looking value — the greppable `marker` proves it is an honest
        "not wired yet", not a mock.
      properties:
        code:
          const: NOT_IMPLEMENTED
          description: Frozen-vocabulary-style code (SCREAMING_SNAKE).
        operationId:
          type: string
          description: The operationId that is not yet wired (e.g. orvexApplyOps).
        marker:
          const: ORVEX_NOT_IMPLEMENTED
          description: The greppable sentinel. Its presence proves the 501 is a real stub, not a mock.

    # ── Framework (Nest/Fastify) error envelope — VERIFIED against the live
    #    engine (curl http://localhost:3000 → {message,error,statusCode}). ─────
    NestError:
      type: object
      description: >-
        The upstream Nest/Fastify error envelope, as observed live from the engine
        (e.g. 404 → {"message":"Workspace not found","error":"Not Found","statusCode":404}).
        Used for framework-level errors (401/404/500). `message` may be a string or
        a string array (class-validator).
      required: [statusCode, error, message]
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string

    # ── Family error envelope — ../orvex-studio-contracts/errors/vocabulary.yaml
    #    (business/domain errors carry a frozen errorCode). ────────────────────
...
    TenantMoveManifest:
      type: object
      additionalProperties: false
      required: [schema_version, tenant_id, stores, s3_prefixes, cursors]
      description: >-
        The versioned manifest describing a single tenant-move (org-move
        generalises to tenant-move: a tenant is a polymorphic user|org). Mirrors
        org-move-manifest.schema.json field-for-field.


        OPEN DECISION (store-list): `stores[]` is coverage-checked against the
        service's DECLARED datastore inventory (cell-contract rule #10, per-repo
        `ORG_MOVE.md`). The engine's declared store inventory for tenant-move is
        NOT yet frozen — TBD-at-delivery (the ORG_MOVE.md coverage check pins it).
      properties:
        schema_version:
          type: integer
          minimum: 1
          description: Manifest schema version; bumped on any breaking change so producers/consumers gate on it.
        tenant_id:
          type: string
          format: uuid
          description: The uuidv7 id of the tenant (user|org) being moved.
        stores:
          type: array
          description: >-
            The relational/logical datastores this service will
            quiesce→export→import→activate for the tenant. OPEN — coverage-checked
            against the engine's declared datastore inventory (TBD-at-delivery).
          items:
            $ref: '#/components/schemas/TenantMoveStore'
        s3_prefixes:
          type: array
          description: The S3 key prefixes owned by this tenant that participate in the move.
          items:
            $ref: '#/components/schemas/TenantMoveS3Prefix'
        cursors:
          type: array
          description: Durable checkpoints (Kafka offsets, outbox positions, external cursors) carried across the move.
          items:
            $ref: '#/components/schemas/TenantMoveCursor'
    TenantMoveStore:
      type: object
      additionalProperties: false
      required: [name, kind]
      properties:
        name:
          type: string
          description: Logical store name (must appear in the engine's declared datastore inventory).
        kind:
          type: string
          enum: [postgres, redis, kafka, s3, external]
          description: Backing store class.
        row_estimate:
          type: integer
          minimum: 0
          description: Optional coarse row estimate for sizing/quiesce planning.
    TenantMoveS3Prefix:
      type: object
      additionalProperties: false
      required: [bucket, prefix]
      properties:
        bucket:
          type: string
          description: Per-cell bucket name (carries the cell).
        prefix:
          type: string
          description: Key prefix, must begin with the tenant_id ({tenant_id}/...).
    TenantMoveCursor:
      type: object
      additionalProperties: false
      required: [source, position]
      properties:
        source:
          type: string
          description: Cursor source (e.g. a topic {domain}-events.{cell}, an outbox table, or an external provider).
        position:
          type: string
          description: Opaque durable position/offset/checkpoint (string to avoid precision loss).
    TenantMoveStepResult:
      type: object
      additionalProperties: false
      required: [tenant_id, step, status]
      description: >-
        INTENDED — coarse-but-real per-step receipt. Delivery fills in the real
        per-step payload (exported byte counts, cursor checkpoints, etc.).
      properties:
        tenant_id:
          type: string
          format: uuid
        step:
          type: string
          enum: [quiesce, export, import, activate]
        status:
          type: string
          enum: [accepted, done]